refactor: move ArgumentError to lib folder
