+++ src/graphql/asLink.ts
@@ -4,7 +4,7 @@ import {
 	LinkField,
 	LinkType
 } from "@prismicio/types/dist/graphql";
-import { ArgumentError } from "../ArgumentError";
+import { ArgumentError } from "../lib/ArgumentError";
 import { LinkResolverFunction } from "./types";
 
 /**
