+++ src/lib/ArgumentError.ts
